+++ src/core/contracts/interfaces/pool/IAlgebraPoolPermissionedActions.sol
@@ -6,9 +6,9 @@ pragma solidity >=0.5.0;
 /// @dev Credit to Uniswap Labs under GPL-2.0-or-later license:
 /// https://github.com/Uniswap/v3-core/tree/main/contracts/interfaces
 interface IAlgebraPoolPermissionedActions {
-  /// @notice Set the community's % share of the fees. Cannot exceed 25% (250). Only factory owner or POOLS_ADMINISTRATOR_ROLE role
+  /// @notice Set the community's % share of the fees. Only factory owner or POOLS_ADMINISTRATOR_ROLE role
   /// @param communityFee new community fee percent in thousandths (1e-3)
-  function setCommunityFee(uint8 communityFee) external;
+  function setCommunityFee(uint16 communityFee) external;
 
   /// @notice Set the new tick spacing values. Only factory owner or POOLS_ADMINISTRATOR_ROLE role
   /// @param newTickSpacing The new tick spacing value
